% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nea-data.r
\docType{data}
\name{can.sig.go}
\alias{can.sig.go}
\title{Example functional gene sets (FGS) from GO and KEGG}
\format{An object, i.e. a list with FGS IDs as names and vectors of gene IDs as member (34 FGSs in total). See see \code{\link{import.gs}}}
\usage{
data(can.sig.go)
}
\description{
Example functional gene sets (FGS) from GO and KEGG
}
\examples{
head(can.sig.go)

}
\keyword{datasets}

