% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2int.r
\name{char2int.fast}
\alias{char2int.fast}
\title{char2int function}
\usage{
char2int.fast(net.list, gs.list.1, gs.list.2 = NULL, Parallelize = 1)
}
\arguments{
\item{net.list}{the global network object, pre-created using \code{\link{import.net}}}

\item{gs.list.1}{AGS or FGS object that lists members of each individual AGS/FGS, pre-created using \code{\link{samples2ags}}}

\item{gs.list.2}{FGS or AGS object that lists members of each individual FGS/AGS, pre-created using \code{\link{samples2ags}}}

\item{Parallelize}{The number of CPU cores to be used for the step "Counting actual network links" (the other steps are sufficiently fast). The option is not supported on Windows.}
}
\description{
This function reads input list objects and maps unique numbers to each member in the network/ genesets.
}
\seealso{
\code{\link{nea.render}}
}
\keyword{internal}
