% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nea-data.r
\docType{data}
\name{fantom5.43samples}
\alias{fantom5.43samples}
\title{A Fantom5 transcriptomes in 43 carincinoma cell samples}
\format{An expression data matrix with the genes as rows and columns as cell line samples}
\usage{
data(fantom5.43samples)
}
\description{
A Fantom5 transcriptomes in 43 carincinoma cell samples
}
\examples{
colnames(fantom5.43samples)

}
\keyword{datasets}

