% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nea-data.r
\docType{data}
\name{net.kegg}
\alias{net.kegg}
\title{A more compact alternative global network for the network enrichment analysis}
\format{A list, see \code{\link{import.net}}}
\usage{
data(net.kegg)
}
\description{
A more compact alternative global network for the network enrichment analysis
}
\examples{
head(net.kegg)

}
\keyword{datasets}

