% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nea-data.r
\docType{data}
\name{tcga.gbm}
\alias{tcga.gbm}
\title{NEArender}
\format{A matrix  with genes as rows (9660 genes with mutations reported in at least one sample) and TCGA samples as columns (291 samples). Wild type (=reference) states are denoted with NA.}
\usage{
data(tcga.gbm)
}
\description{
TCGA point mutations from the glioblastoma multiforme cohort
}
\examples{
table(is.na(tcga.gbm))
}
\keyword{datasets}

